+++ src/initctl.c
@@ -339,7 +339,7 @@ int main(int argc, char *argv[]) {
         log_set_target(LOG_TARGET_SYSLOG_OR_KMSG);
         log_parse_environment();
 
-        log_info("systemd-initctl running as pid %llu", (unsigned long long) getpid());
+        log_info("systemd-initctl running as pid %lu", (unsigned long) getpid());
 
         if ((n = sd_listen_fds(true)) < 0) {
                 log_error("Failed to read listening file descriptors from environment: %s", strerror(-r));
@@ -387,7 +387,7 @@ fail:
 
         server_done(&server);
 
-        log_info("systemd-initctl stopped as pid %llu", (unsigned long long) getpid());
+        log_info("systemd-initctl stopped as pid %lu", (unsigned long) getpid());
 
         dbus_shutdown();
