+++ src/main.c
@@ -116,7 +116,7 @@ _noreturn_ static void crash(int sig) {
                         else if (!WCOREDUMP(status))
                                 log_error("Caught <%s>, core dump failed.", strsignal(sig));
                         else
-                                log_error("Caught <%s>, dumped core as pid %llu.", strsignal(sig), (unsigned long long) pid);
+                                log_error("Caught <%s>, dumped core as pid %lu.", strsignal(sig), (unsigned long) pid);
                 }
         }
 
@@ -152,7 +152,7 @@ _noreturn_ static void crash(int sig) {
                         _exit(1);
                 }
 
-                log_info("Successfully spawned crash shall as pid %llu.", (unsigned long long) pid);
+                log_info("Successfully spawned crash shall as pid %lu.", (unsigned long) pid);
         }
 
         log_info("Freezing execution.");
